fix warning
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@24709 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/krb5/test_gic.c
@@ -42,7 +42,7 @@ static krb5_error_code
 lr_proc(krb5_context context, krb5_last_req_entry **e, void *ctx)
 {
     while (e && *e) {
-	printf("e type: %d value: %d\n", (*e)->lr_type, (*e)->value);
+	printf("e type: %d value: %d\n", (*e)->lr_type, (int)(*e)->value);
 	e++;
     }
     return 0;
